pyproject.toml
[project]
name = "viral_usher"
version = "0.1.0"
description = "Command-line pipeline for viral data processing"
authors = [{ name = "Angie Hinrichs", email = "hinrichs@ucsc.edu" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["requests", "biopython"]

[project.scripts]
viral_usher = "viral_usher.cli:main"